chore: cleanup path utils
