+++ packages/sdk/src/utils.ts
@@ -27,3 +27,28 @@ export const makeBip44Path = (
   const { account, change, index } = bip44Path;
   return `m/44'/${coinType}'/${account}'/${change}'/${index}'`;
 };
+
+/**
+ * Return a properly formatted Sapling path
+ * @param {number} coinType - SLIP-044 Coin designation
+ * @param {number} account - numbered from index 
+ in sequentially increasing manner. Defined as in BIP 44
+ * @returns {string} Sapling path
+ */
+export const makeSaplingPath = (coinType: number, account: number): string => {
+  return `m/32'/${coinType}'/${account}'`;
+};
+
+/**
+ * Return a properly formatted Sapling path array
+ * @param {number} coinType - SLIP-044 Coin designation
+ * @param {number} account - numbered from index 
+ in sequentially increasing manner. Defined as in BIP 44
+ * @returns {string} Sapling path array
+ */
+export const makeSaplingPathArray = (
+  coinType: number,
+  account: number
+): Uint32Array => {
+  return new Uint32Array([32, coinType, account]);
+};
